move driverjoystick into controllers
The driver joystick is closer to a controller than a util class
